python/2020-SDK/messaging/send-mms/requirements.txt
flask
bandwidth-sdk
requests
